functions/requirements.txt
azure-functions
azure-functions-durable
azure-storage-blob
azure-cosmos
azure-identity
azure-keyvault-secrets
requests
hashlib
uuid
